+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.3.0] - 2021-09-03
+
 #### Features
 
 - Encoder harness
@@ -30,7 +32,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.2.0] - 2021-08-16
 
 <!-- next-url -->
-[Unreleased]: https://github.com/epage/git-stack/compare/v0.2.5...HEAD
+[Unreleased]: https://github.com/epage/git-stack/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/epage/git-stack/compare/v0.2.5...v0.3.0
 [0.2.5]: https://github.com/epage/git-stack/compare/v0.2.4...v0.2.5
 [0.2.4]: https://github.com/epage/git-stack/compare/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/epage/git-stack/compare/v0.2.2...v0.2.3